Working on changing ownership of an object.

+++ Semester 2/CommonNetworking/AnimPacket.h
@@ -4,7 +4,7 @@
 
 struct AnimData : PacketData
 {
-	int _state;
+	float _state;
 };
